refactor: unified env.mjs handling across apps
