+++ apps/governance.mento.org/app/api/contract/services/blockchain-explorer-service.ts
@@ -1,4 +1,4 @@
-import { env } from "../../../env.mjs";
+import { env } from "@/env.mjs";
 
 export type BlockchainExplorerSource = "blockscout" | "celoscan";
