+++ apps/reserve.mento.org/sentry.edge.config.ts
@@ -4,7 +4,7 @@
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/
 
 import * as Sentry from "@sentry/nextjs";
-import { env } from "./env.mjs";
+import { env } from "@/env.mjs";
 
 Sentry.init({
   dsn: env.NEXT_PUBLIC_SENTRY_DSN_RESERVE,
