+++ apps/ui.mento.org/next.config.ts
@@ -1,5 +1,5 @@
 import type { NextConfig } from "next";
-import { env } from "./env.mjs";
+import { env } from "@/env.mjs";
 
 const nextConfig: NextConfig = {
   /* Performance Optimizations */
